refactor(nodes): move file content I/O nodes to lib.os namespace
The four remaining file-content nodes belonged with the other filesystem
nodes, not in a dedicated workspace namespace. Move them into lib-os.ts
and rename their node types:

- nodetool.workspace.ReadTextFile   -> lib.os.ReadTextFile
- nodetool.workspace.WriteTextFile  -> lib.os.WriteTextFile
- nodetool.workspace.ReadBinaryFile -> lib.os.ReadBinaryFile
- nodetool.workspace.WriteBinaryFile-> lib.os.WriteBinaryFile

The workspace module is now empty and removed, along with WORKSPACE_NODES.
Updates exports, generated DSL bindings, docs (moved to nodes/lib/os),
the graph-planner baseline namespaces, and tests.
