--- docs/nodes/nodetool/workspace/index.md
@@ -1,13 +1,0 @@
----
-layout: page
-title: "nodetool.workspace Nodes"
----
-
-This namespace contains 4 node(s).
-
-## Available Nodes
-
-- **[Read Binary File](readbinaryfile.md)** - Read a binary file from the workspace as base64-encoded string.
-- **[Read Text File](readtextfile.md)** - Read a text file from the workspace.
-- **[Write Binary File](writebinaryfile.md)** - Write binary data (base64-encoded) to a file in the workspace.
-- **[Write Text File](writetextfile.md)** - Write text to a file in the workspace.
