+++ packages/base-nodes/src/index.ts
@@ -154,13 +154,6 @@ export {
 export { OutputNode, PreviewNode, OUTPUT_NODES } from "@nodetool-ai/audio-nodes/nodes/output";
 export { WorkflowNode, WORKFLOW_NODES } from "@nodetool-ai/core-nodes/nodes/workflow";
 export { SubgraphNode, SUBGRAPH_NODES } from "@nodetool-ai/core-nodes/nodes/subgraph";
-export {
-  ReadTextFileNode,
-  WriteTextFileNode,
-  ReadBinaryFileNode,
-  WriteBinaryFileNode,
-  WORKSPACE_NODES
-} from "@nodetool-ai/automation-nodes/nodes/workspace";
 export { CompareImagesNode, COMPARE_NODES } from "@nodetool-ai/core-nodes/nodes/compare";
 export {
   LoadDocumentFileNode,
@@ -406,6 +399,10 @@ export {
 } from "@nodetool-ai/video-nodes/nodes/model3d";
 export {
   WorkspaceDirectoryLibNode,
+  ReadTextFileNode,
+  WriteTextFileNode,
+  ReadBinaryFileNode,
+  WriteBinaryFileNode,
   OpenWorkspaceDirectoryLibNode,
   FileExistsLibNode,
   ListFilesLibNode,
@@ -727,7 +724,6 @@ import { INPUT_NODES } from "@nodetool-ai/core-nodes/nodes/input";
 import { OUTPUT_NODES } from "@nodetool-ai/audio-nodes/nodes/output";
 import { WORKFLOW_NODES } from "@nodetool-ai/core-nodes/nodes/workflow";
 import { SUBGRAPH_NODES } from "@nodetool-ai/core-nodes/nodes/subgraph";
-import { WORKSPACE_NODES } from "@nodetool-ai/automation-nodes/nodes/workspace";
 import { COMPARE_NODES } from "@nodetool-ai/core-nodes/nodes/compare";
 import { DOCUMENT_NODES } from "@nodetool-ai/document-nodes/nodes/document";
 import { DATA_NODES } from "@nodetool-ai/data-nodes/nodes/data";
@@ -811,7 +807,6 @@ export const ALL_BASE_NODES: readonly NodeClass[] = [
   ...OUTPUT_NODES,
   ...WORKFLOW_NODES,
   ...SUBGRAPH_NODES,
-  ...WORKSPACE_NODES,
   ...COMPARE_NODES,
   ...DOCUMENT_NODES,
   ...DATA_NODES,
